refactor: clean up, modularize
