+++ ObservableWebView/Toolbar/UrlSearchBarTextField.swift
@@ -1,0 +1,35 @@
+//
+//  UrlSearchBarTextField.swift
+//  ObservableWebView
+//
+//  Created by Justin Bush on 3/19/24.
+//
+
+import SwiftUI
+
+struct UrlSearchBarTextField: View {
+  let manager: ObservableWebViewManager
+  @State private var text: String = ""
+  
+  var body: some View {
+    TextField("Search or type URL", text: $text)
+      .textFieldStyle(RoundedBorderTextFieldStyle())
+      .onSubmit {
+        manager.load(text)
+      }
+      .onChange(of: manager.urlString) {
+        observedUrlChange()
+      }
+  }
+  
+  func observedUrlChange() {
+    guard let urlString = manager.urlString else { return }
+    text = urlString
+  }
+}
+
+
+#Preview {
+  ContentView()
+    .frame(width: 400, height: 600)
+}
